pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "BrainNormativeRealNVP"
version = "0.1.0"
description = "A library for running a conditional RealNVP-based normative model"
readme = "README.md"
authors = [{ name = "Mai Ho", email = "mai.ho@unsw.edu.au" }]
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "optuna==3.4.0",
    "statsmodels==0.13.2",
    "matplotlib==3.5.3",
    "seaborn>=0.11.2",
    "numpy==1.23.2",
    "torch==1.12.1",
    "pandas==1.4.3",
    "scikit-learn==1.3.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.65.0",
    "openpyxl>=3.1.5"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
brain-nvp-train = "scripts.train_model:main"
brain-nvp-inference = "scripts.run_inference:main"
